style: Add ContextIndicator component to AgStyle
This commit adds a new component called ContextIndicator to the AgStyle file in the portal-administration application. The ContextIndicator component is a styled span element that displays a colored circle based on the value of the 'active' prop. If the prop is set to 'true', the background color of the circle is set to the warning color defined in the design tokens. Otherwise, the background color is set to 'none'. This component is used for indicating the active state of a context in the application.