feat: Add isGlobal and isContextual properties to PortalApp and PortalApplication types
This commit adds the properties isGlobal and isContextual to the PortalApp and PortalApplication types in the index.ts file. These properties indicate whether an app or application is global or contextual. The isGlobal property is added to the PortalApp type, while both properties are added to the PortalApplication type. This change allows for better categorization and management of apps and applications in the portal-administration application.