chore: Update package.json and add new dependency
Bump the version in package.json from 0.0.5 to 0.0.6. Also, add the new dependency "@equinor/fusion-framework-module-event" to the project.

refactor: Update use-navigate-on-context-change.ts

Refactor the use-navigate-on-context-change.ts file to use the updated modules from the fusion framework. Remove the unused imports and update the event listener to use the "onCurrentContextChanged" event.